+++ frontend/src/metabase/lib/formatting/field.ts
@@ -1,0 +1,7 @@
+export function formatField(field) {
+  if (!field) {
+    return "";
+  }
+
+  return field.dimensions?.name || field.display_name || field.name;
+}
