+++ frontend/src/metabase/lib/formatting/sql.ts
@@ -1,0 +1,14 @@
+export function formatSQL(sql) {
+  if (typeof sql === "string") {
+    sql = sql.replace(/\sFROM/, "\nFROM");
+    sql = sql.replace(/\sLEFT JOIN/, "\nLEFT JOIN");
+    sql = sql.replace(/\sWHERE/, "\nWHERE");
+    sql = sql.replace(/\sGROUP BY/, "\nGROUP BY");
+    sql = sql.replace(/\sORDER BY/, "\nORDER BY");
+    sql = sql.replace(/\sLIMIT/, "\nLIMIT");
+    sql = sql.replace(/\sAND\s/, "\n   AND ");
+    sql = sql.replace(/\sOR\s/, "\n    OR ");
+
+    return sql;
+  }
+}
